UBERF-8313: Fix select workspace sorting
